% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMADA.summary.R
\name{ARMADA.summary}
\alias{ARMADA.summary}
\title{Scores of the covariates X}
\usage{
ARMADA.summary(X, resultat.ARMADA.select)
}
\arguments{
\item{X}{the matrix (or data.frame) of covariates, dimension n*p (n is the sample size, p the number of covariates). X must have colnames.}

\item{resultat.ARMADA.select}{the output of the ARMADA.select function: a list  with 8 vectors, called: genes_rf_thres, genes_rf_interp, genes_lasso, genes_bonferroni, genes_BH, genes_qvalues, genes_localfdr, genes_FAMT. The 8 vectors are the selected covariates by the corresponding selection methods.}
}
\value{
gene_list: data.frame with p rows and 2 columns, the first column gives the names of the covariates, the second column is the scores of the covariates.
}
\description{
Scores of the covariates X
}
\details{
The function ARMADA.summary gives the scores of all the covariates. The score of a variable is an integer between 0 and 8, and represents the number of selections of this variable by the 8 selection methods.
}
